# backend/docs/swagger.yaml
basePath: /
definitions:
...
  controller.InterestPatchInput:
    properties:
      name:
        type: string
    type: object
...
  models.Artist:
    properties:
      concerts:
        items:
          $ref: '#/definitions/models.Concert'
        type: array
      createdAt:
        type: string
      deletedAt:
        type: string
      id:
        description: gorm.Model
        type: string
      interest:
        $ref: '#/definitions/models.Interest'
      interestId:
        type: integer
      name:
        type: string
      updatedAt:
        type: string
    type: object
  models.Category:
    properties:
      concertCategories:
        items:
          $ref: '#/definitions/models.ConcertCategory'
        type: array
      createdAt:
        type: string
      deletedAt:
        type: string
      id:
        description: gorm.Model
        type: integer
      name:
        type: string
      updatedAt:
        type: string
    type: object
  models.Concert:
    properties:
      artist:
        $ref: '#/definitions/models.Artist'
      artistId:
        type: string
      concertCategories:
        items:
          $ref: '#/definitions/models.ConcertCategory'
        type: array
      createdAt:
        type: string
      date:
        type: string
      deletedAt:
        type: string
      description:
        type: string
      id:
        description: gorm.Model
        type: string
      image:
        type: string
      interests:
        items:
          $ref: '#/definitions/models.Interest'
        type: array
      location:
        type: string
      name:
        type: string
      organization:
        $ref: '#/definitions/models.Organization'
      organizationId:
        type: string
      updatedAt:
        type: string
    type: object
  models.ConcertCategory:
    properties:
      availableTickets:
        type: integer
      category:
        $ref: '#/definitions/models.Category'
      categoryId:
        type: integer
      concert:
        allOf:
        - $ref: '#/definitions/models.Concert'
        description: Tickets          []Ticket `gorm:"-"`
      concertId:
        type: string
      createdAt:
        type: string
      deletedAt:
        type: string
      id:
        description: gorm.Model
        type: string
      price:
        type: number
      soldTickets:
        type: integer
      tickets:
        items:
          $ref: '#/definitions/models.Ticket'
        type: array
      updatedAt:
        type: string
    type: object
  models.Conversation:
    properties:
      buyer:
        $ref: '#/definitions/models.User'
      buyerId:
        type: string
      createdAt:
        type: string
      deletedAt:
        type: string
      id:
        description: gorm.Model
        type: string
      messages:
        items:
          $ref: '#/definitions/models.Message'
        type: array
      price:
        type: number
      seller:
        $ref: '#/definitions/models.User'
      sellerId:
        type: string
      ticketListing:
        $ref: '#/definitions/models.TicketListing'
      ticketListingId:
        type: string
      updatedAt:
        type: string
    type: object
  models.Interest:
    properties:
      concerts:
        items:
          $ref: '#/definitions/models.Concert'
        type: array
      createdAt:
        type: string
      deletedAt:
        type: string
      id:
        description: gorm.Model
        type: integer
      name:
        type: string
      updatedAt:
        type: string
      users:
        items:
          $ref: '#/definitions/models.User'
        type: array
    type: object
  models.Message:
    properties:
      author:
        $ref: '#/definitions/models.User'
      authorId:
        type: string
      content:
        type: string
      conversation:
        $ref: '#/definitions/models.Conversation'
      conversationId:
        type: string
      createdAt:
        type: string
      deletedAt:
        type: string
      id:
        description: gorm.Model
        type: string
      readed:
        type: boolean
      sentAt:
        type: string
      updatedAt:
        type: string
    type: object
  models.Organization:
    properties:
      concerts:
        items:
          $ref: '#/definitions/models.Concert'
        type: array
      createdAt:
        type: string
      deletedAt:
        type: string
      description:
        type: string
      id:
        description: gorm.Model
        type: string
      image:
        type: string
      name:
        type: string
      updatedAt:
        type: string
      users:
        items:
          $ref: '#/definitions/models.User'
        type: array
    type: object
  models.Sale:
    properties:
      buyer:
        $ref: '#/definitions/models.User'
      buyerId:
        type: string
      createdAt:
        type: string
      deletedAt:
        type: string
      finalPrice:
        type: number
      id:
        description: gorm.Model
        type: string
      seller:
        $ref: '#/definitions/models.User'
      sellerId:
        type: string
      ticketListingId:
        type: string
      ticketSold:
        $ref: '#/definitions/models.TicketListing'
      updatedAt:
        type: string
    type: object
  models.Ticket:
    properties:
      concertCategory:
        $ref: '#/definitions/models.ConcertCategory'
      concertCategoryId:
        type: string
      createdAt:
        type: string
      deletedAt:
        type: string
      id:
        description: gorm.Model
        type: string
      maxPrice:
        type: number
      ticketListings:
        items:
          $ref: '#/definitions/models.TicketListing'
        type: array
      updatedAt:
        type: string
      user:
        $ref: '#/definitions/models.User'
      userId:
        type: string
    type: object
  models.TicketListing:
    properties:
      conversations:
        items:
          $ref: '#/definitions/models.Conversation'
        type: array
      createdAt:
        type: string
      deletedAt:
        type: string
      id:
        description: |-
          gorm.Model
          ID            uuid.UUID       `gorm:"unique;type:uuid;primaryKey;default:uuid_generate_v4()" json:"id"`
          Price         float64         `gorm:"not null" json:"price"`
          Status        string          `gorm:"not null" json:"status"`
          CreatedAt     time.Time       `json:"createdAt"`
          UpdatedAt     time.Time       `json:"updatedAt"`
          DeletedAt     *time.Time      `gorm:"index"`
          TicketId      uuid.UUID       `gorm:"not null" json:"ticketId"`
        type: string
      price:
        type: number
      sale:
        $ref: '#/definitions/models.Sale'
      status:
        type: string
      ticket:
        $ref: '#/definitions/models.Ticket'
      ticketId:
        type: string
      updatedAt:
        type: string
    type: object
  models.User:
    properties:
      conversationsAsBuyer:
        items:
          $ref: '#/definitions/models.Conversation'
        type: array
      conversationsAsSeller:
        items:
          $ref: '#/definitions/models.Conversation'
        type: array
      createdAt:
        type: string
      deletedAt:
        type: string
      email:
        description: |-
          l'email de l'utilisateur

          required: true
          example: john@doe.com
        type: string
      firstname:
        description: |-
          le prénom de l'utilisateur

          required: true
          example: John
        type: string
      id:
        description: |-
          gorm.Model
          l'ID de l'utilisateur

          required: true
          example: 123e4567-e89b-12d3-a456-426614174000
        type: string
      image:
        type: string
      interests:
        items:
          $ref: '#/definitions/models.Interest'
        type: array
      lastConnexion:
        type: string
      lastname:
        description: |-
          le nom de l'utilisateur

          required: true
          example: Doe
        type: string
      messages:
        items:
          $ref: '#/definitions/models.Message'
        type: array
      organization:
        $ref: '#/definitions/models.Organization'
      organizationId:
        type: string
      resetCodeExpiration:
        type: string
      role:
        type: string
      salesAsBuyer:
        items:
          $ref: '#/definitions/models.Sale'
        type: array
      salesAsSeller:
        items:
          $ref: '#/definitions/models.Sale'
        type: array
      tickets:
        items:
          $ref: '#/definitions/models.Ticket'
        type: array
      updatedAt:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is the swagger documentation for the Weezemaster API.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Weezemaster API
  version: "1.0"
paths:
...
  /interests:
    get:
      description: Récupère tous les centres d'intérêts
      operationId: get-all-interests
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Interest'
            type: array
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - Bearer: []
      summary: Récupère tous les centres d'intérêts
      tags:
      - Interests
    post:
      description: Créé un centre d'intérêt
      operationId: create-interest
      parameters:
      - description: Centre d'intérêt à créer
        in: body
        name: interest
        required: true
        schema:
          $ref: '#/definitions/models.Interest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            items:
              $ref: '#/definitions/models.Interest'
            type: array
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - Bearer: []
      summary: Créé un centre d'intérêt
      tags:
      - Interests
  /interests/{id}:
    delete:
      description: Supprime un centre d'intérêt par ID
      operationId: delete-interest
      parameters:
      - description: ID du centre d'intérêt
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - Bearer: []
      summary: Supprime un centre d'intérêt
      tags:
      - Interests
    get:
      description: Récupère un centre d'intérêt par ID
      operationId: get-interest
      parameters:
      - description: ID du centre d'intérêt
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Interest'
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - Bearer: []
      summary: Récupère un centre d'intérêt
      tags:
      - Interests
    patch:
      description: Modifie un centre d'intérêt par ID
      operationId: update-interest
      parameters:
      - description: ID du centre d'intérêt
        in: path
        name: id
        required: true
        type: integer
      - description: Centre d'intérêt à modifier
        in: body
        name: interest
        required: true
        schema:
          $ref: '#/definitions/controller.InterestPatchInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Interest'
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - Bearer: []
      summary: Modifie un centre d'intérêt
      tags:
      - Interests
...
  /messages:
    post:
      description: Créer un message
      operationId: post-message
      parameters:
      - description: Contenu du message
        in: body
        name: content
        required: true
        schema:
          type: string
      - description: Message
        format: uuid
        in: body
        name: conversation_id
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Message'
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - Bearer: []
      summary: Créer un message
      tags:
      - Messages
...
  /user/interests:
    get:
      description: Récupère les centres d'intérêt de l'utilisateur
      operationId: get-user-interests
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Interest'
            type: array
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - Bearer: []
      summary: Récupère les centres d'intérêt de l'utilisateur
      tags:
      - Interests
  /user/interests/{id}:
    delete:
      description: Supprime un centre d'intérêt de l'utilisateur
      operationId: remove-user-interest
      parameters:
      - description: ID du centre d'intérêt
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - Bearer: []
      summary: Supprime un centre d'intérêt de l'utilisateur
      tags:
      - Interests
    post:
      description: Ajoute un centre d'intérêt à l'utilisateur
      operationId: add-user-interest
      parameters:
      - description: ID du centre d'intérêt
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Interest'
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - Bearer: []
      summary: Ajoute un centre d'intérêt à l'utilisateur
      tags:
      - Interests
...
  /users/{id}:
    delete:
      description: Supprime un utilisateur par ID
      operationId: delete-user
      parameters:
      - description: ID de l'utilisateur
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - Bearer: []
      summary: Supprime un utilisateur
      tags:
      - Users
    get:
      description: Récupère un utilisateur par ID
      operationId: get-user
      parameters:
      - description: ID de l'utilisateur
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema: {}
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - Bearer: []
      summary: Récupère un utilisateur
      tags:
      - Users
    patch:
      description: Modifie un utilisateur par ID
      operationId: update-user
      parameters:
      - description: ID de l'utilisateur
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Requête de modification d'utilisateur
        in: body
        name: formData
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema: {}
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - Bearer: []
      summary: Modifie un utilisateur
      tags:
      - Users
...
securityDefinitions:
  Bearer:
    description: 'Provide your JWT token in the format: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"